move RTMAllModelPackInfoCrashCallable
