+++ src/main/java/com/anatawa12/fixRtm/crash/RTMAllModelPackInfoCrashCallable.kt
@@ -1,6 +1,7 @@
-package com.anatawa12.fixRtm
+package com.anatawa12.fixRtm.crash
 
 import com.anatawa12.fixRtm.rtm.modelpack.modelset.dummies.IFixRTMDummyModelSet
+import com.anatawa12.fixRtm.state
 import com.anatawa12.fixRtm.utils.Base64OutputStream
 import com.anatawa12.fixRtm.utils.ToStringOutputStream
 import com.google.gson.GsonBuilder
